# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RelayLab Coordinator API",
    "version": "1.2.0",
    "description": "The browser-facing REST contract. Coordinator, JSON-RPC, repository, and database details remain private implementation concerns. Group extension: opt-in local demo sessions and independent run-review decisions. Demo identities are selectable by anyone and are not production authentication."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local coordinator"
    }
  ],
  "paths": {
    "/api/experiments": {
      "get": {
        "operationId": "listExperiments",
        "summary": "List saved experiments newest first",
        "responses": {
          "200": {
            "description": "Saved experiments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Experiment"
                  }
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/PersistenceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createExperiment",
        "summary": "Validate and create an experiment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Experiment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Experiment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid experiment input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/PersistenceUnavailable"
          }
        }
      }
    },
    "/api/experiments/{experimentId}": {
      "get": {
        "operationId": "getExperiment",
        "summary": "Get one experiment and its run history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExperimentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment with durable run history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetails"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ExperimentNotFound"
          },
          "503": {
            "$ref": "#/components/responses/PersistenceUnavailable"
          }
        }
      },
      "delete": {
        "summary": "Delete one experiment and, by cascade, its runs",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExperimentId"
          }
        ],
        "responses": {
          "204": {
            "description": "The experiment and its runs were deleted"
          },
          "404": {
            "description": "No experiment with that identifier"
          },
          "503": {
            "description": "Persistence temporarily unavailable"
          },
          "409": {
            "description": "Duplicate submission, already-decided review or protected evidence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experimentId}/runs": {
      "post": {
        "operationId": "runExperiment",
        "summary": "Execute and persist one distributed run",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExperimentId"
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Optional exchange id minted by the caller; becomes the JSON-RPC id and is echoed in X-Correlation-Id"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "description": "Optional key for a retry of the same logical run. A settled run is replayed with 200; an unsettled attempt is sent again and the dependency dedupes"
          }
        ],
        "responses": {
          "200": {
            "description": "A settled run replayed for a repeated Idempotency-Key; X-Idempotent-Replay: true"
          },
          "201": {
            "description": "Run persisted, including controlled dependency failures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRun"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ExperimentNotFound"
          },
          "503": {
            "$ref": "#/components/responses/PersistenceUnavailable"
          }
        }
      }
    },
...
    "/api/runs/{runId}/execution": {
      "get": {
        "operationId": "getRunnerExecution",
        "summary": "Inspect the runner-owned execution for a coordinator receipt without changing the recorded attempt",
        "description": "Shared local lab read, like experiment history. Makes a real GetExecution gRPC call; no access to the runner database. Cache-Control: no-store. Legacy JSON-RPC receipts have no gRPC execution.",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerExecution"
                }
              }
            }
          },
          "404": {
            "description": "Run or runner operation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Run has no configured gRPC execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Runner unavailable, deadline exceeded or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ExperimentId": {
        "name": "experimentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      }
    },
    "responses": {
      "ExperimentNotFound": {
        "description": "Experiment does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PersistenceUnavailable": {
        "description": "The configured relational store is temporarily unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "ExperimentBehavior": {
        "type": "string",
        "enum": [
          "healthy",
          "slow",
          "unavailable",
          "malformed"
        ]
      },
      "ExperimentInput": {
        "type": "object",
        "required": [
          "name",
          "behavior",
          "payload"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "behavior": {
            "$ref": "#/components/schemas/ExperimentBehavior"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "Experiment": {
        "type": "object",
        "description": "V1 clients depend on these required fields and tolerate additional optional fields.",
        "required": [
          "id",
          "name",
          "behavior",
          "payload",
          "createdAt"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "type": "string"
          },
          "behavior": {
            "$ref": "#/components/schemas/ExperimentBehavior"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RunOutcome": {
        "type": "string",
        "enum": [
          "success",
          "downstream_error",
          "timeout",
          "invalid_response",
          "unreachable"
        ]
      },
      "ExperimentRun": {
        "type": "object",
        "required": [
          "id",
          "experimentId",
          "outcome",
          "httpStatus",
          "durationMs",
          "response",
          "createdAt"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "experimentId": {
            "type": "integer",
            "minimum": 1
          },
          "outcome": {
            "$ref": "#/components/schemas/RunOutcome"
          },
          "httpStatus": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 100,
            "maximum": 599
          },
          "rpcErrorCode": {
            "description": "The dependency's JSON-RPC error code when its method failed, otherwise null",
            "type": [
              "integer",
              "null"
            ]
          },
          "idempotencyKey": {
            "type": [
              "string",
              "null"
            ],
            "description": "The caller's Idempotency-Key for this attempt, if any"
          },
          "durationMs": {
            "type": "integer",
            "minimum": 1
          },
          "response": {
            "description": "The preserved dependency envelope or null when no response arrived",
            "type": [
              "object",
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ExperimentDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Experiment"
          },
          {
            "type": "object",
            "required": [
              "runs"
            ],
            "properties": {
              "runs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ExperimentRun"
                }
              }
            }
          }
        ]
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        ]
      },
...
      "RunnerExecution": {
        "type": "object",
        "required": [
          "executionId",
          "operationId",
          "experimentRef",
          "state",
          "outcome",
          "resultJson",
          "startedAt",
          "completedAt"
        ],
        "properties": {
          "executionId": {
            "type": "string",
            "format": "uuid"
          },
          "operationId": {
            "type": "string"
          },
          "experimentRef": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "INTERRUPTED"
            ]
          },
          "outcome": {
            "type": "string"
          },
          "resultJson": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "description": "Empty while running; otherwise ISO timestamp."
          }
        }
      }
    },
    "securitySchemes": {
      "DemoSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token from POST /api/demo-sessions. Selectable local demo accounts only; sessions expire after one hour or coordinator restart."
      }
    }
  }
}